reorganized files into srce folder
